@@ -7,28 +7,51 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "test": "jest",
-    "eject": "expo eject"
+    "test": "jest src/views && jest src/components",
+    "t": "jest",
+    "eject": "expo eject",
+    "lint": "eslint --ext .tsx --ext .ts src/ && prettier --write ./src"
   },
   "dependencies": {
-    "@expo/vector-icons": "^10.0.0",
+    "@expo/vector-icons": "^12.0.0",
+    "@react-native-async-storage/async-storage": "~1.15.0",
+    "@react-native-community/datetimepicker": "3.5.2",
     "@react-native-community/masked-view": "0.1.10",
+    "@react-native-picker/picker": "1.16.1",
     "@react-navigation/native": "^5.7.3",
     "@react-navigation/stack": "^5.9.0",
     "@types/styled-components": "^5.1.3",
     "@types/styled-components-react-native": "^5.1.1",
+    "@types/styled-system": "^5.1.10",
+    "@types/uuid": "^8.3.0",
+    "@use-expo/font": "^2.0.0",
     "axios": "latest",
-    "expo": "~38.0.8",
-    "expo-status-bar": "^1.0.2",
-    "react": "~16.11.0",
-    "react-dom": "~16.11.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
+    "expo": "^42.0.0",
+    "expo-app-loading": "^1.1.2",
+    "expo-camera": "~11.2.1",
+    "expo-font": "~9.2.1",
+    "expo-linear-gradient": "~9.2.0",
+    "expo-media-library": "~12.1.2",
+    "expo-permissions": "~12.1.1",
+    "expo-status-bar": "~1.0.4",
+    "formik": "^2.2.6",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
     "react-native-animatable": "latest",
-    "react-native-gesture-handler": "~1.6.0",
-    "react-native-reanimated": "~1.9.0",
-    "react-native-safe-area-context": "~3.0.7",
-    "react-native-screens": "~2.9.0",
-    "react-native-web": "~0.11.7",
+    "react-native-app-intro-slider": "^4.0.4",
+    "react-native-gesture-handler": "~1.10.2",
+    "react-native-image-zoom-viewer": "^3.0.1",
+    "react-native-masked-text": "^1.13.0",
+    "react-native-modal-datetime-picker": "^9.2.0",
+    "react-native-reanimated": "~2.2.0",
+    "react-native-safe-area-context": "3.2.0",
+    "react-native-screens": "~3.4.0",
+    "react-native-toast-message": "^1.4.9",
+    "react-native-uuid": "^2.0.1",
+    "react-native-web": "~0.13.12",
     "react-navigation": "^4.4.0",
     "react-redux": "latest",
     "reactotron-react-native": "^5.0.0",
@@ -36,18 +59,27 @@
     "redux-promise": "latest",
     "redux-saga": "~1.1.1",
     "redux-thunk": "~2.3.0",
-    "styled-components": "latest"
+    "styled-components": "latest",
+    "styled-system": "^5.1.5",
+    "yup": "^0.32.9"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.6",
+    "@babel/core": "~7.9.0",
     "@testing-library/react-native": "^7.1.0",
     "@types/jest": "^26.0.15",
-    "@types/react": "~16.9.41",
-    "@types/react-native": "~0.62.13",
-    "babel-preset-expo": "^7.0.0",
-    "jest-expo": "^39.0.0",
+    "@types/react": "~16.9.35",
+    "@types/react-native": "~0.63.2",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
+    "babel-plugin-module-resolver": "^4.1.0",
+    "babel-preset-expo": "8.3.0",
+    "eslint": "^7.22.0",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "jest-expo": "^42.0.0",
     "plop": "latest",
-    "typescript": "~3.9.5"
+    "prettier": "^2.2.1",
+    "typescript": "~4.0.0"
   },
   "jest": {
     "preset": "jest-expo"
