@@ -1,10 +1,15 @@
-import React from 'react';
-import { InitialProvider, ModalProvider } from '../../hooks';
+import * as React from 'react';
+import { ThemeProvider } from '~/hooks/Theme';
+import { EventProvider, DateProvider, AlertProvider } from '~/hooks';
 
 export const GlobalProvider: React.FC = ({ children }) => {
   return (
-    <ModalProvider>
-      <InitialProvider>{children}</InitialProvider>
-    </ModalProvider>
+    <ThemeProvider>
+      <AlertProvider>
+        <DateProvider>
+          <EventProvider>{children}</EventProvider>
+        </DateProvider>
+      </AlertProvider>
+    </ThemeProvider>
   );
 };
