@@ -11,21 +11,21 @@ const ComponentDeTest: React.FC<IComponentDeTest> = ({ press }) => {
   const [password, setPassword] = useState('');
 
   return (
-    <View testID={`ComponentDeTest`}>
+    <View testID={'ComponentDeTest'}>
       <TextInput
-        testID={`email`}
+        testID={'email'}
         value={email}
         onChangeText={setEmail}
         placeholder={'Email'}
       />
       <TextInput
-        testID={`password`}
+        testID={'password'}
         value={password}
         onChangeText={setPassword}
         placeholder={'Password'}
       />
       <Button
-        testID={`button`}
+        testID={'button'}
         onPress={() => press(email, password)}
         title={'Botão'}
       />
@@ -52,7 +52,7 @@ const exemploDeFn = jest.fn();
 describe('Render ComponentDeTest', () => {
   it('Should be ComponentDeTest exist', () => {
     const { getByTestId } = render(<ComponentDeTest press={exemploDeFn} />);
-    const currentElement = getByTestId(`ComponentDeTest`);
+    const currentElement = getByTestId('ComponentDeTest');
     expect(currentElement).toBeTruthy();
   });
 });
@@ -62,9 +62,9 @@ describe('ComponentDeTest ', () => {
   it('Should be ComponentDeTest is functional', () => {
     // capturando elementos
     const { getByTestId } = render(<ComponentDeTest press={exemploDeFn} />);
-    const emailInput = getByTestId(`email`);
-    const passwordInput = getByTestId(`password`);
-    const buttonElement = getByTestId(`button`);
+    const emailInput = getByTestId('email');
+    const passwordInput = getByTestId('password');
+    const buttonElement = getByTestId('button');
     // checkando se os elementos estão na tela
     expect(emailInput).toBeTruthy();
     expect(passwordInput).toBeTruthy();
